pyproject.toml
[tool.poetry]
name = "pass-operator"
version = "0.0.1"
description = "A kubernetes operator that syncs and decrypts secrets from pass git repositories"
authors = ["Emma Doyle <emma@premiscale.com>"]
maintainers = ["Emma Doyle <emma@premiscale.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
packages = [
  { include = "passoperator", from = "src" }
]
include = [
  "LICENSE"
]
keywords = [
    "python",
    "kubernetes",
    "secrets",
    "operator",
    "pass"
]

[tool.poetry.dependencies]
python = "^3.10"
kubernetes = "^29.0.0"
kopf = "^1.36.2"
python-gnupg = "^0.5.2"
gitpython = "^3.1.41"
dacite = "^1.8.1"
pyhumps = "^3.8.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"
poetry = "^1.8.3"
mypy = "^1.0.1"
pylint = "^3.0.3"
pytest = "^7.4.4"
coverage = "^7.4.1"
deepdiff = "^6.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
passoperator = "passoperator.daemon:main"

[tool.mypy]
python_version = "3.10"
strict_optional = "true"
ignore_missing_imports = "true"

[tool.coverage.run]
# omit = [".*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 60

[tool.pytest.ini_options]
addopts = "--junitxml=./junit_test_results.xml"
testpaths = [
  "tests/src",
  "tests/e2e"
]

[tool.pylint."MESSAGES CONTROL"]
disable = "invalid-name,wrong-import-order,superfluous-parens,line-too-long,missing-final-newline,logging-fstring-interpolation,too-many-instance-attributes,unused-argument,unused-import"
fail-under = 9

[tool.pylint.MASTER]
init-hook = "import sys; sys.path.append('.')"